fix: address critical and high severity code review findings
Critical fixes:
- tbft_checkpoint_region: replace first-wins digest logic with multi-candidate
  tracking (TBFT_CERT_MAX_VALS slots) to prevent Byzantine bias
- tbft_replica: verify HMAC authenticator on incoming Pre_prepare before storing
- tbft_replica: sign outgoing View_change with RSA; verify signature on receipt
- tbft_replica: remove self-id spoofing path in handle_checkpoint; store own
  checkpoint at send time in execute_committed instead
- tbft_replica: RSA-sign Reply messages; verify signatures in Byz_recv_reply
- tbft_principal: replace PKCS#1 v1.5 key-wrap with OAEP(SHA-256)
- tbft_principal: enforce TBFT_ANTI_REPLAY_WINDOW_US time-window check, not
  just monotonicity
- tbft_partition: replace XOR-then-hash with concat-then-hash (PSA multi-part
  SHA-256) for collision-resistant Merkle parent digests

High severity fixes:
- tbft_principal: cache persistent PSA HMAC key handles to avoid import/destroy
  on every hot-path MAC operation; use psa_mac_verify for timing-safe checks
- tbft_node: add sender_id bounds check in tbft_node_verify_auth and
  tbft_node_verify_sig before indexing principals[]
- tbft_replica: fix execute_committed loop bound (last_prepared+1 → last_prepared)
- tbft_replica: build Reply in r->out_buf, not a stack VLA
- tbft_transport_udp: accumulate sent bytes in broadcast loop (sent = r → sent += r)
- tbft_transport_espnow: increment next_msg_id in single-fragment unicast path;
  fix oversized stack buffer (TBFT_MAX_MESSAGE_SIZE → ESPNOW_MAX_DATA_LEN)
- tbft_libbyz: add lower-bound check in Byz_modify1; free principals on
  setup_principals failure; zero sig bytes if signing skipped; drop view
  requirement in reply matching (match on rid only)
- tbft_partition/state: change tbft_part_t.version and update_leaf() to
  tbft_seqno_t to eliminate int32_t truncation of 64-bit sequence numbers

+++ src/tbft_state.c
@@ -146,7 +146,7 @@ void tbft_state_checkpoint(tbft_state_t *state, tbft_seqno_t seqno)
         const uint8_t *block = state->mem + (size_t)i * TBFT_BLOCK_SIZE;
         tbft_msg_digest(block, TBFT_BLOCK_SIZE, &state->block_digests[i]);
         tbft_ptree_update_leaf(&state->ptree, i, &state->block_digests[i],
-                               (int32_t)seqno);
+                               seqno);
     }
 
     /* Store the checkpoint record */
@@ -183,7 +183,7 @@ tbft_seqno_t tbft_state_rollback(tbft_state_t *state)
                         TBFT_BLOCK_SIZE, &state->block_digests[bidx]);
         tbft_ptree_update_leaf(&state->ptree, bidx,
                                &state->block_digests[bidx],
-                               (int32_t)rec->seqno);
+                               rec->seqno);
     }
 
     cow_zero(state);
